requirements-dev.txt
python-dotenv==1.0.0
pytest==7.2.0